Added mysql session provider to do the db work for the connection service, also changed the login service interface to expose an api to return the session key.

+++ src/login/login_service.cc
@@ -322,7 +322,6 @@ void LoginService::SendEnumerateCharacterId_(std::shared_ptr<LoginClient> login_
     
     login_client->session->SendMessage(character_message);
 }
-tuple<string, uint64_t> LoginService::GetSessionKey(uint64_t session_id)
-{
+tuple<string, uint32_t> LoginService::GetSessionKey(uint32_t session_id) {
     return make_tuple("", 1);
 }
